+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Added `first` and `single` assertion for `Iterable`
+- Added sequence assertions to mirror iterable
 
 ### Fixed
 - Fixed iterable assertions that take a block that could have multiple assertions.